Hi Mohsen,
I hope this message finds you well. I've been working with the project and encountered an issue related to the CircularIndeterminateProgressBar while scrolling down the RecipeListScreen. After investigating, I believe the problem lies within the animations of the CircularProgressIndicator, resulting in a NoSuchMethodError during the loading stage.

I've made some adjustments that seem to resolve this issue and wanted to reach out to discuss . If it's alright with you, I'd be happy to share the details or submit a pull request with the changes for your review.

Your project is clean and enjoyable to work with overall, and I'd like to contribute to resolving this particular issue.

Looking forward to your guidance or feedback on this matter.

Best regards
https://t.me/AliIsazadh
isazadhali@gmail.com